backend_requirement.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.14.0
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.1
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1